backend/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Uploads==0.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pandas==2.2.3
openpyxl==3.1.5
werkzeug==3.0.4
